Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
cloudinary==1.38.0
stripe==7.11.0
paypalrestsdk==1.13.1
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
argon2-cffi==23.1.0









